isehan-master/requirements.txt
Django==2.2.9
mysql-connector-python==8.0.18
graphene-django==2.8.0
django-filter==2.2.0
Pillow==7.2.0
django-graphql-jwt==0.3.1
